requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram-media-group==0.5.1
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
openai==0.27.4
pydantic==1.10.7
regex==2023.3.23
requests==2.28.2
tiktoken==0.3.3
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.3
yarl==1.8.2
python-keycloak==3.3.0